requirements.txt
langchain
python-docx